added env variable in instrumentation-client.ts
