--- instrumentation.ts
@@ -1,16 +1,0 @@
-import * as Sentry from '@sentry/nextjs';
-
-import { version, name } from './package.json';
-
-export function register() {
-  Sentry.init({
-    dsn: process.env.NEXT_PUBLIC_SENTRY_DSN,
-    tracesSampleRate: 1,
-    debug: false,
-    environment: process.env.NEXT_PUBLIC_SENTRY_ENV,
-    enabled: process.env.NODE_ENV === 'production',
-    release: `${name}@${version}`,
-  });
-}
-
-export const onRequestError = Sentry.captureRequestError;
